Fix multiple processing loops
Turns out threads still run even when the app is
onPause.  Returning from onPause calls onResume which
would spawn a new processing loop.

Now we use a boolean, running, that is set to true
in doit() (called from onResume()), set to false in
onPause(), and the processing loop quits when running
is false.