Dim the display at night
Tapping on the light bulb icon in the toolbar will toggle the
main screen between bright and dark.  When the screen is dark
then tapping anywhere on the screen will turn it back on.

Also, the screen automatically goes dark at a set time and goes
back on at a different set time.