Redesign get_frame
It now has two modes:

names - returns all the image names for a given list
get - returns the specific image name requested

Also, the E (end) meta data now has a generation number as it's
string.  The generation number is monotonically increased everytime
there is a change to the image files for a list so that the Android
app can know when it's list of image names is out of date.