Avoid stale images in the grid view
The ImageAdapter recycles views as the user scrolls.  Since getting
a thumbnail from the image server can take a while this means the
following sequence can happen for a particular grid entry:

1) adapter expands view, asks for image A
2) user scrolls so adapter expands same view again, asks for image B
3) image B comes back first
4) image A comes back, overriding the correct image B

Solve the problem by setting the position into the views tag.
The thumbnail code then checks to make sure the position its trying
to set matches the tag in the view and just doesn't update the
view if the positions differ.